Move wasm specific methods separate modules
